+++ ClassWork/src/shapes/Rectangle.java
@@ -6,10 +6,10 @@ package shapes;
  */
 public class Rectangle implements Shape{
 	
-    private int length;
-    private int width;
+    private double length;
+    private double width;
 
-    public Rectangle(int length, int width) {
+    public Rectangle(double length, double width) {
         this.length = length;
         this.width = width;
     }
